Conserto da importação das libs referentes ao JPAStrategy.
Atividades Realizadas
- Executar manualmente o mesmo experimento para o hibernate e o eclipselink, ver a diferenças entre as tecnologias JPA. (curiosidade)
- Testar, no retorno de um SELECT, se uma lista é retornada com um tipo pojo? Utilizar o instanceOf para isso.(Feito, com getClass)

+++ trunk/jpameter/src/jpa/compiler/Compiler.java
@@ -255,9 +255,11 @@ public class Compiler {
 			for(File file: files){
 				if(!file.isDirectory()){
 					String name = file.getName().toUpperCase();
-					String ext = name.substring(name.lastIndexOf("."), name.length()); 
-					if(ext.equals(".SF")||ext.equals(".DSA")||ext.equals(".RSA")||ext.equals(".EC"))
-						file.delete();
+					if(name.lastIndexOf(".") != -1){
+						String ext = name.substring(name.lastIndexOf("."), name.length());
+						if(ext.equals(".SF")||ext.equals(".DSA")||ext.equals(".RSA")||ext.equals(".EC"))
+							file.delete();
+					}
 				}
 			}
 		}
